Improve frontend structure
