+++ frontend/src/components/App/App.test.tsx
@@ -1,0 +1,8 @@
+import { render, screen } from "@testing-library/react";
+import App from ".";
+
+test("renders learn react link", () => {
+  render(<App />);
+  const linkElement = screen.getByText(/learn react/i);
+  expect(linkElement).toBeInTheDocument();
+});
