+++ frontend/src/components/App/index.tsx
@@ -1,0 +1,21 @@
+import React from "react";
+import logo from "logo.svg";
+import "./App.css";
+
+const App: React.FC = (): JSX.Element => {
+  return (
+    <div className="App">
+      <header className="App-header">
+        <img src={logo} className="App-logo" alt="logo" />
+        <p>
+          Edit <code>src/components/App/index.tsx</code> and save to reload.
+        </p>
+        <a className="App-link" href="https://reactjs.org" target="_blank" rel="noopener noreferrer">
+          Learn React
+        </a>
+      </header>
+    </div>
+  );
+};
+
+export default App;
